+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     * Etc.
 * Changed existing fax properties to give a dictionary of the properties they actually contain rather than just the number. This makes them behave like the newly added email properties.
 * Fixed issues with `Task` properties being incorrect.
+* Added implementation for PtypErrorCode.
 
 **v0.34.3**
 * Fixed issue that may have caused other olefile types to raise the wrong type of error when passed to `openMsg`.